refactor #194
