+++ src/nexus/core/nexus_fs.py
@@ -2,12 +2,10 @@
 
 from __future__ import annotations
 
-import asyncio
 import builtins
 import contextlib
-import fnmatch
 import hashlib
-import re
+import json
 import threading
 from datetime import UTC, datetime
 from pathlib import Path
@@ -16,7 +14,7 @@ from typing import Any
 from sqlalchemy import select
 
 from nexus.backends.backend import Backend
-from nexus.core.exceptions import ConflictError, InvalidPathError, NexusFileNotFoundError
+from nexus.core.exceptions import InvalidPathError, NexusFileNotFoundError
 from nexus.core.export_import import (
     CollisionDetail,
     ExportFilter,
@@ -25,6 +23,10 @@ from nexus.core.export_import import (
 )
 from nexus.core.filesystem import NexusFilesystem
 from nexus.core.metadata import FileMetadata
+from nexus.core.nexus_fs_core import NexusFSCoreMixin
+from nexus.core.nexus_fs_permissions import NexusFSPermissionsMixin
+from nexus.core.nexus_fs_search import NexusFSSearchMixin
+from nexus.core.nexus_fs_versions import NexusFSVersionsMixin
 from nexus.core.permissions import OperationContext, Permission
 from nexus.core.router import NamespaceConfig, PathRouter
 from nexus.parsers import MarkItDownParser, ParserRegistry
@@ -32,7 +34,13 @@ from nexus.parsers.types import ParseResult
 from nexus.storage.metadata_store import SQLAlchemyMetadataStore
 
 
-class NexusFS(NexusFilesystem):
+class NexusFS(
+    NexusFSCoreMixin,
+    NexusFSSearchMixin,
+    NexusFSPermissionsMixin,
+    NexusFSVersionsMixin,
+    NexusFilesystem,
+):
     """
     Unified filesystem for Nexus.
 
@@ -267,18 +275,6 @@ class NexusFS(NexusFilesystem):
 
         return path
 
-    def _compute_etag(self, content: bytes) -> str:
-        """
-        Compute ETag for file content.
-
-        Args:
-            content: File content
-
-        Returns:
-            ETag (MD5 hash)
-        """
-        return hashlib.md5(content).hexdigest()
-
     def _get_parent_path(self, path: str) -> str | None:
         """
         Get parent directory path from a file path.
@@ -420,808 +416,6 @@ class NexusFS(NexusFilesystem):
         )
 
         self.metadata.put(metadata)
-
-    def read(
-        self, path: str, context: OperationContext | None = None, return_metadata: bool = False
-    ) -> bytes | dict[str, Any]:
-        """
-        Read file content as bytes.
-
-        Args:
-            path: Virtual path to read
-            context: Optional operation context for permission checks (uses default if not provided)
-            return_metadata: If True, return dict with content and metadata (etag, version, modified_at).
-                           If False, return only content bytes (default: False)
-
-        Returns:
-            If return_metadata=False: File content as bytes
-            If return_metadata=True: Dict with keys:
-                - content: File content as bytes
-                - etag: Content hash (SHA-256) for optimistic concurrency
-                - version: Current version number
-                - modified_at: Last modification timestamp
-                - size: File size in bytes
-
-        Raises:
-            NexusFileNotFoundError: If file doesn't exist
-            InvalidPathError: If path is invalid
-            BackendError: If read operation fails
-            AccessDeniedError: If access is denied based on tenant isolation
-            PermissionError: If user doesn't have read permission
-
-        Examples:
-            >>> # Read content only
-            >>> content = nx.read("/workspace/data.json")
-            >>> print(content)
-            b'{"key": "value"}'
-
-            >>> # Read with metadata for optimistic concurrency
-            >>> result = nx.read("/workspace/data.json", return_metadata=True)
-            >>> content = result['content']
-            >>> etag = result['etag']
-            >>> # Later, write with version check
-            >>> nx.write("/workspace/data.json", new_content, if_match=etag)
-        """
-        path = self._validate_path(path)
-
-        # Check read permission (v0.3.0)
-        self._check_permission(path, Permission.READ, context)
-
-        # Route to backend with access control
-        route = self.router.route(
-            path,
-            tenant_id=self.tenant_id,
-            agent_id=self.agent_id,
-            is_admin=self.is_admin,
-            check_write=False,
-        )
-
-        # Check if file exists in metadata
-        meta = self.metadata.get(path)
-        if meta is None or meta.etag is None:
-            raise NexusFileNotFoundError(path)
-
-        # Read from routed backend using content hash
-        content = route.backend.read_content(meta.etag)
-
-        # Return content with metadata if requested
-        if return_metadata:
-            return {
-                "content": content,
-                "etag": meta.etag,
-                "version": meta.version,
-                "modified_at": meta.modified_at,
-                "size": meta.size,
-            }
-
-        return content
-
-    def write(
-        self,
-        path: str,
-        content: bytes,
-        context: OperationContext | None = None,
-        if_match: str | None = None,
-        if_none_match: bool = False,
-        force: bool = False,
-    ) -> dict[str, Any]:
-        """
-        Write content to a file with optional optimistic concurrency control.
-
-        Creates parent directories if needed. Overwrites existing files.
-        Updates metadata store.
-
-        Automatically deduplicates content using CAS.
-
-        Args:
-            path: Virtual path to write
-            content: File content as bytes
-            context: Optional operation context for permission checks (uses default if not provided)
-            if_match: Optional etag for optimistic concurrency control (v0.3.9).
-                     If provided, write only succeeds if current file etag matches this value.
-                     Prevents concurrent modification conflicts.
-            if_none_match: If True, write only if file doesn't exist (create-only mode)
-            force: If True, skip version check and overwrite unconditionally (dangerous!)
-
-        Returns:
-            Dict with metadata about the written file:
-                - etag: Content hash (SHA-256) of the written content
-                - version: New version number
-                - modified_at: Modification timestamp
-                - size: File size in bytes
-
-        Raises:
-            InvalidPathError: If path is invalid
-            BackendError: If write operation fails
-            AccessDeniedError: If access is denied (tenant isolation or read-only namespace)
-            PermissionError: If path is read-only or user doesn't have write permission
-            ConflictError: If if_match is provided and doesn't match current etag (v0.3.9)
-            FileExistsError: If if_none_match=True and file already exists
-
-        Examples:
-            >>> # Simple write (no version checking)
-            >>> result = nx.write("/workspace/data.json", b'{"key": "value"}')
-            >>> print(result['etag'], result['version'])
-
-            >>> # Optimistic concurrency control
-            >>> result = nx.read("/workspace/data.json", return_metadata=True)
-            >>> new_content = modify(result['content'])
-            >>> try:
-            ...     nx.write("/workspace/data.json", new_content, if_match=result['etag'])
-            ... except ConflictError:
-            ...     print("File was modified by another agent!")
-
-            >>> # Create-only mode
-            >>> nx.write("/workspace/new.txt", b'content', if_none_match=True)
-        """
-        path = self._validate_path(path)
-
-        # Route to backend with write access check FIRST (to check tenant/agent isolation)
-        # This must happen before permission check so AccessDeniedError is raised before PermissionError
-        route = self.router.route(
-            path,
-            tenant_id=self.tenant_id,
-            agent_id=self.agent_id,
-            is_admin=self.is_admin,
-            check_write=True,
-        )
-
-        # Check if path is read-only
-        if route.readonly:
-            raise PermissionError(f"Path is read-only: {path}")
-
-        # Get existing metadata for permission check and update detection (single query)
-        now = datetime.now(UTC)
-        meta = self.metadata.get(path)
-
-        # Check write permission (v0.3.0)
-        # Only check permissions if the file is owned by the current user
-        # This allows namespace routing to override Unix permissions when needed
-        # Rationale: Namespace isolation is PRIMARY, Unix permissions are SECONDARY
-        if meta is not None:
-            ctx = context or self._default_context
-
-            # Only check permissions if we own the file
-            # If someone else owns it but the router allows write access, namespace wins
-            if meta.owner == ctx.user:
-                # Existing file owned by us - check permissions to prevent accidental overwrites
-                self._check_permission(path, Permission.WRITE, context)
-            # If file is owned by someone else, skip permission check
-            # The router has already validated namespace-level access (tenant/agent isolation)
-        # NOTE: For new files, we do NOT check parent directory permissions because:
-        # 1. The router has already validated namespace-level access (tenant/agent isolation)
-        # 2. The new file will get correct owner/group/mode via permission policies
-        # 3. Checking parent permissions can cause false rejections when namespaces
-        #    allow access but parent was created by a different user
-
-        # Optimistic concurrency control (v0.3.9)
-        if not force:
-            # Check if_none_match (create-only mode)
-            if if_none_match and meta is not None:
-                raise FileExistsError(f"File already exists: {path}")
-
-            # Check if_match (version check)
-            if if_match is not None:
-                if meta is None:
-                    # File doesn't exist, can't match etag
-                    raise ConflictError(
-                        path=path,
-                        expected_etag=if_match,
-                        current_etag="(file does not exist)",
-                    )
-                elif meta.etag != if_match:
-                    # Version mismatch - conflict detected!
-                    raise ConflictError(
-                        path=path,
-                        expected_etag=if_match,
-                        current_etag=meta.etag or "(no etag)",
-                    )
-
-        # Write to routed backend - returns content hash
-        content_hash = route.backend.write_content(content)
-
-        # NOTE: Do NOT delete old content when updating a file!
-        # Version history (v0.3.5) preserves references to old content hashes.
-        # Old content should only be deleted when ALL versions are deleted.
-        # CAS reference counting handles cleanup automatically.
-
-        # Apply permission policy for new files, preserve for existing files
-        # Also apply policy if existing file has no permissions set (migration case)
-        if meta is None or (meta.owner is None and meta.group is None and meta.mode is None):
-            # New file or existing file without permissions - try policy first, then inheritance
-            owner = None
-            group = None
-            mode = None
-
-            # Build context for variable substitution in policy
-            policy_context = {
-                "agent_id": self.agent_id or "unknown",
-                "tenant_id": self.tenant_id or "default",
-                "user_id": self.agent_id or "unknown",
-            }
-
-            # Try permission policy first
-            policy_result = self.policy_matcher.apply_policy(
-                path=path,
-                tenant_id=self.tenant_id,
-                context=policy_context,
-                is_directory=False,
-            )
-
-            if policy_result:
-                owner, group, mode = policy_result
-            else:
-                # No policy matched - fall back to parent directory inheritance
-                owner, group, mode = self._inherit_permissions_from_parent(path, is_directory=False)
-        else:  # Existing file with permissions - preserve them
-            owner = meta.owner
-            group = meta.group
-            mode = meta.mode
-
-        # Calculate new version number (increment if updating)
-        new_version = (meta.version + 1) if meta else 1
-
-        # Store metadata with content hash as both etag and physical_path
-        metadata = FileMetadata(
-            path=path,
-            backend_name=self.backend.name,
-            physical_path=content_hash,  # CAS: hash is the "physical" location
-            size=len(content),
-            etag=content_hash,  # SHA-256 hash for integrity
-            created_at=meta.created_at if meta else now,
-            modified_at=now,
-            version=new_version,
-            owner=owner,  # Apply policy or inherit from parent
-            group=group,  # Apply policy or inherit from parent
-            mode=mode,  # Apply policy or inherit from parent
-        )
-
-        self.metadata.put(metadata)
-
-        # Auto-parse file if enabled and format is supported
-        if self.auto_parse:
-            self._auto_parse_file(path)
-
-        # Return metadata for optimistic concurrency control (v0.3.9)
-        return {
-            "etag": content_hash,
-            "version": new_version,
-            "modified_at": now,
-            "size": len(content),
-        }
-
-    def _auto_parse_file(self, path: str) -> None:
-        """Auto-parse a file in the background (fire-and-forget).
-
-        Args:
-            path: Virtual path to the file
-        """
-        try:
-            # Check if parser is available for this file type
-            self.parser_registry.get_parser(path)
-
-            # Run parsing in a background thread (fire-and-forget)
-            thread = threading.Thread(
-                target=self._parse_in_thread,
-                args=(path,),
-                daemon=True,
-            )
-            # Track thread for graceful shutdown
-            with self._parser_threads_lock:
-                self._parser_threads.append(thread)
-            thread.start()
-        except Exception:
-            # Silently ignore if no parser available or parsing fails
-            pass
-
-    def _parse_in_thread(self, path: str) -> None:
-        """Parse file in a background thread.
-
-        Args:
-            path: Virtual path to the file
-        """
-        # Silently ignore parsing errors
-        with contextlib.suppress(Exception):
-            # Run async parse in a new event loop (thread-safe)
-            asyncio.run(self.parse(path, store_result=True))
-
-    def delete(self, path: str, context: OperationContext | None = None) -> None:
-        """
-        Delete a file.
-
-        Removes file from backend and metadata store.
-        Decrements reference count in CAS (only deletes when ref_count=0).
-
-        Args:
-            path: Virtual path to delete
-            context: Optional operation context for permission checks (uses default if not provided)
-
-        Raises:
-            NexusFileNotFoundError: If file doesn't exist
-            InvalidPathError: If path is invalid
-            BackendError: If delete operation fails
-            AccessDeniedError: If access is denied (tenant isolation or read-only namespace)
-            PermissionError: If path is read-only or user doesn't have write permission
-        """
-        path = self._validate_path(path)
-
-        # Route to backend with write access check FIRST (to check tenant/agent isolation)
-        # This must happen before permission check so AccessDeniedError is raised before PermissionError
-        route = self.router.route(
-            path,
-            tenant_id=self.tenant_id,
-            agent_id=self.agent_id,
-            is_admin=self.is_admin,
-            check_write=True,
-        )
-
-        # Check if path is read-only
-        if route.readonly:
-            raise PermissionError(f"Cannot delete from read-only path: {path}")
-
-        # Check if file exists in metadata
-        meta = self.metadata.get(path)
-        if meta is None:
-            raise NexusFileNotFoundError(path)
-
-        # Check write permission for delete (v0.3.0)
-        # This comes AFTER tenant isolation check so AccessDeniedError takes precedence
-        self._check_permission(path, Permission.WRITE, context)
-
-        # Delete from routed backend CAS (decrements ref count)
-        if meta.etag:
-            route.backend.delete_content(meta.etag)
-
-        # Remove from metadata
-        self.metadata.delete(path)
-
-    def rename(self, old_path: str, new_path: str) -> None:
-        """
-        Rename/move a file by updating its path in metadata.
-
-        This is a metadata-only operation that does NOT copy file content.
-        The file's content remains in the same location in CAS storage,
-        only the virtual path is updated in the metadata database.
-
-        This makes rename/move operations instant, regardless of file size.
-
-        Args:
-            old_path: Current virtual path
-            new_path: New virtual path
-
-        Raises:
-            NexusFileNotFoundError: If source file doesn't exist
-            FileExistsError: If destination path already exists
-            InvalidPathError: If either path is invalid
-            PermissionError: If either path is read-only
-            AccessDeniedError: If access is denied (tenant isolation)
-
-        Example:
-            >>> nx.rename('/workspace/old.txt', '/workspace/new.txt')
-            >>> nx.rename('/folder-a/file.txt', '/shared/folder-a/file.txt')
-        """
-        old_path = self._validate_path(old_path)
-        new_path = self._validate_path(new_path)
-
-        # Route both paths
-        old_route = self.router.route(
-            old_path,
-            tenant_id=self.tenant_id,
-            agent_id=self.agent_id,
-            is_admin=self.is_admin,
-            check_write=True,  # Need write access to source
-        )
-        new_route = self.router.route(
-            new_path,
-            tenant_id=self.tenant_id,
-            agent_id=self.agent_id,
-            is_admin=self.is_admin,
-            check_write=True,  # Need write access to destination
-        )
-
-        # Check if paths are read-only
-        if old_route.readonly:
-            raise PermissionError(f"Cannot rename from read-only path: {old_path}")
-        if new_route.readonly:
-            raise PermissionError(f"Cannot rename to read-only path: {new_path}")
-
-        # Check if source exists
-        if not self.metadata.exists(old_path):
-            raise NexusFileNotFoundError(old_path)
-
-        # Check if destination already exists
-        if self.metadata.exists(new_path):
-            raise FileExistsError(f"Destination path already exists: {new_path}")
-
-        # Perform metadata-only rename (no CAS I/O!)
-        self.metadata.rename_path(old_path, new_path)
-
-    def exists(self, path: str) -> bool:
-        """
-        Check if a file or directory exists.
-
-        Args:
-            path: Virtual path to check
-
-        Returns:
-            True if file or implicit directory exists, False otherwise
-        """
-        try:
-            path = self._validate_path(path)
-            # Check if file exists explicitly
-            if self.metadata.exists(path):
-                return True
-            # Check if it's an implicit directory (has files beneath it)
-            return self.metadata.is_implicit_directory(path)
-        except InvalidPathError:
-            return False
-
-    def list(
-        self,
-        path: str = "/",
-        recursive: bool = True,
-        details: bool = False,
-        prefix: str | None = None,
-        context: OperationContext | None = None,
-    ) -> builtins.list[str] | builtins.list[dict[str, Any]]:
-        """
-        List files in a directory.
-
-        Args:
-            path: Directory path to list (default: "/")
-            recursive: If True, list all files recursively; if False, list only direct children (default: True)
-            details: If True, return detailed metadata; if False, return paths only (default: False)
-            prefix: (Deprecated) Path prefix to filter by - for backward compatibility.
-                    When used, lists all files recursively with this prefix.
-            context: Optional operation context for permission filtering (uses default if not provided)
-
-        Returns:
-            List of file paths (if details=False) or list of file metadata dicts (if details=True).
-            Each metadata dict contains: path, size, modified_at, etag
-            Results are filtered by read permission.
-
-        Examples:
-            # List all files recursively (default)
-            fs.list()  # Returns: ["/file1.txt", "/dir/file2.txt", "/dir/subdir/file3.txt"]
-
-            # List files in root directory only (non-recursive)
-            fs.list("/", recursive=False)  # Returns: ["/file1.txt"]
-
-            # List files recursively with details
-            fs.list(details=True)  # Returns: [{"path": "/file1.txt", "size": 100, ...}, ...]
-
-            # Old API (deprecated but supported)
-            fs.list(prefix="/dir")  # Returns all files under /dir recursively
-        """
-        # Handle backward compatibility with old 'prefix' parameter
-        if prefix is not None:
-            # Old API: list(prefix="/path") - always recursive
-            if prefix:
-                prefix = self._validate_path(prefix)
-            all_files = self.metadata.list(prefix)
-            results = all_files
-        else:
-            # New API: list(path="/", recursive=False)
-            if path:
-                path = self._validate_path(path)
-
-            # Ensure path ends with / for directory listing
-            if not path.endswith("/"):
-                path = path + "/"
-
-            # Get all files with this prefix
-            all_files = self.metadata.list(path if path != "/" else "")
-
-            if recursive:
-                # Include all files under this path
-                results = all_files
-            else:
-                # Only include files directly in this directory (no subdirectories)
-                results = []
-                for meta in all_files:
-                    # Remove the prefix to get relative path
-                    rel_path = meta.path[len(path) :] if path != "/" else meta.path[1:]
-                    # If there's no "/" in the relative path, it's in this directory
-                    if "/" not in rel_path:
-                        results.append(meta)
-
-        # Filter by read permission (v0.3.0)
-        if self._enforce_permissions:
-            ctx = context or self._default_context
-            result_paths = [meta.path for meta in results]
-            allowed_paths = self._permission_enforcer.filter_list(result_paths, ctx)
-            # Filter results to only include allowed paths
-            results = [meta for meta in results if meta.path in allowed_paths]
-
-        # Sort by path name
-        results.sort(key=lambda m: m.path)
-
-        # Add directories to results (infer from file paths + check backend)
-        # This ensures empty directories show up in listings
-        directories = set()
-
-        if not recursive:
-            # For non-recursive listings, infer immediate subdirectories from file paths
-            base_path = path if path != "/" else ""
-
-            # Get all files to infer directories
-            all_files_for_dirs = self.metadata.list(base_path)
-            for meta in all_files_for_dirs:
-                # Get relative path
-                rel_path = meta.path[len(path) :] if path != "/" else meta.path[1:]
-                # Check if there's a directory component
-                if "/" in rel_path:
-                    # Extract first directory component
-                    dir_name = rel_path.split("/")[0]
-                    dir_path = path + dir_name if path != "/" else "/" + dir_name
-                    directories.add(dir_path)
-
-            # Check backend for empty directories (directories with no files)
-            # This catches newly created directories using the helper method
-            backend_dirs = self._get_backend_directory_entries(path)
-            directories.update(backend_dirs)
-
-        if details:
-            file_results = [
-                {
-                    "path": meta.path,
-                    "size": meta.size,
-                    "modified_at": meta.modified_at,
-                    "created_at": meta.created_at,
-                    "etag": meta.etag,
-                    "mime_type": meta.mime_type,
-                    "is_directory": False,
-                }
-                for meta in results
-            ]
-
-            # Add directory entries
-            dir_results = [
-                {
-                    "path": dir_path,
-                    "size": 0,
-                    "modified_at": None,
-                    "created_at": None,
-                    "etag": None,
-                    "mime_type": None,
-                    "is_directory": True,
-                }
-                for dir_path in sorted(directories)
-            ]
-
-            # Combine and sort
-            all_results = file_results + dir_results
-            all_results.sort(key=lambda x: str(x["path"]))
-            return all_results
-        else:
-            # Return paths only
-            all_paths = [meta.path for meta in results] + sorted(directories)
-            all_paths.sort()
-            return all_paths
-
-    def glob(self, pattern: str, path: str = "/") -> builtins.list[str]:
-        """
-        Find files matching a glob pattern.
-
-        Supports standard glob patterns:
-        - `*` matches any sequence of characters (except `/`)
-        - `**` matches any sequence of characters including `/` (recursive)
-        - `?` matches any single character
-        - `[...]` matches any character in the brackets
-
-        Args:
-            pattern: Glob pattern to match (e.g., "**/*.py", "data/*.csv", "test_*.py")
-            path: Base path to search from (default: "/")
-
-        Returns:
-            List of matching file paths, sorted by name
-
-        Examples:
-            # Find all Python files recursively
-            fs.glob("**/*.py")  # Returns: ["/src/main.py", "/tests/test_foo.py", ...]
-
-            # Find all CSV files in data directory
-            fs.glob("*.csv", "/data")  # Returns: ["/data/file1.csv", "/data/file2.csv"]
-
-            # Find all test files
-            fs.glob("test_*.py")  # Returns: ["/test_foo.py", "/test_bar.py"]
-        """
-        if path:
-            path = self._validate_path(path)
-
-        # Get all files
-        all_files = self.metadata.list("")
-
-        # Build full pattern
-        if not path.endswith("/"):
-            path = path + "/"
-        if path == "/":
-            full_pattern = pattern
-        else:
-            # Remove leading / from path for pattern matching
-            base_path = path[1:] if path.startswith("/") else path
-            full_pattern = base_path + pattern
-
-        # Match files against pattern
-        # Handle ** for recursive matching
-        if "**" in full_pattern:
-            # Convert glob pattern to regex
-            # Split by ** to handle recursive matching
-            parts = full_pattern.split("**")
-
-            regex_parts = []
-            for i, part in enumerate(parts):
-                if i > 0:
-                    # ** matches zero or more path segments
-                    # This can be empty or ".../", so use (?:.*/)? for optional match
-                    regex_parts.append("(?:.*/)?")
-
-                # Escape and convert wildcards in this part
-                escaped = re.escape(part)
-                escaped = escaped.replace(r"\*", "[^/]*")
-                escaped = escaped.replace(r"\?", ".")
-                escaped = escaped.replace(r"\[", "[").replace(r"\]", "]")
-
-                # Remove leading / from all parts since it's handled by ** or the anchor
-                # Note: re.escape() doesn't escape /, so we check for it directly
-                while escaped.startswith("/"):
-                    escaped = escaped[1:]
-
-                regex_parts.append(escaped)
-
-            regex_pattern = "^/" + "".join(regex_parts) + "$"
-
-            matches = []
-            for meta in all_files:
-                if re.match(regex_pattern, meta.path):
-                    matches.append(meta.path)
-        else:
-            # Use fnmatch for simpler patterns
-            matches = []
-            for meta in all_files:
-                # Remove leading / for matching
-                file_path = meta.path[1:] if meta.path.startswith("/") else meta.path
-                if fnmatch.fnmatch(file_path, full_pattern):
-                    matches.append(meta.path)
-
-        return sorted(matches)
-
-    def grep(
-        self,
-        pattern: str,
-        path: str = "/",
-        file_pattern: str | None = None,
-        ignore_case: bool = False,
-        max_results: int = 1000,
-        search_mode: str = "auto",
-    ) -> builtins.list[dict[str, Any]]:
-        r"""
-        Search file contents using regex patterns.
-
-        Args:
-            pattern: Regex pattern to search for in file contents
-            path: Base path to search from (default: "/")
-            file_pattern: Optional glob pattern to filter files (e.g., "*.py")
-            ignore_case: If True, perform case-insensitive search (default: False)
-            max_results: Maximum number of results to return (default: 1000)
-            search_mode: Content search mode (default: "auto")
-                - "auto": Try parsed text first, fallback to raw (default)
-                - "parsed": Only search parsed text (skip files without parsed content)
-                - "raw": Only search raw file content (skip parsing)
-
-        Returns:
-            List of match dicts, each containing:
-            - file: File path
-            - line: Line number (1-indexed)
-            - content: Matched line content
-            - match: The matched text
-            - source: Source type - "parsed" or "raw"
-
-        Examples:
-            # Search for "TODO" in all files (auto mode - tries parsed first)
-            fs.grep("TODO")
-            # Returns: [{"file": "/main.py", "line": 42, "content": "...", "source": "raw"}, ...]
-
-            # Search for function definitions in Python files
-            fs.grep(r"def \w+", file_pattern="**/*.py")
-
-            # Search only parsed text from PDFs
-            fs.grep("revenue", file_pattern="**/*.pdf", search_mode="parsed")
-
-            # Search only raw content (skip parsing)
-            fs.grep("TODO", search_mode="raw")
-
-            # Case-insensitive search
-            fs.grep("error", ignore_case=True)
-        """
-        if path:
-            path = self._validate_path(path)
-
-        # Validate search_mode
-        valid_modes = {"auto", "parsed", "raw"}
-        if search_mode not in valid_modes:
-            raise ValueError(
-                f"Invalid search_mode: {search_mode}. Must be one of: {', '.join(valid_modes)}"
-            )
-
-        # Compile regex pattern
-        flags = re.IGNORECASE if ignore_case else 0
-        try:
-            regex = re.compile(pattern, flags)
-        except re.error as e:
-            raise ValueError(f"Invalid regex pattern: {e}") from e
-
-        # Get files to search
-        files: list[str]
-        if file_pattern:
-            files = self.glob(file_pattern, path)
-        else:
-            # Get all files under path
-            if not path.endswith("/"):
-                path = path + "/"
-            prefix = path if path != "/" else ""
-            all_files = self.metadata.list(prefix)
-            files = [meta.path for meta in all_files]
-
-        # Search through files
-        results: list[dict[str, Any]] = []
-        for file_path in files:
-            if len(results) >= max_results:
-                break
-
-            try:
-                text: str | None = None
-                source: str = "raw"
-
-                # Get parsed text if needed
-                if search_mode in ("auto", "parsed"):
-                    parsed_text = self.metadata.get_file_metadata(file_path, "parsed_text")
-                    if parsed_text:
-                        text = parsed_text
-                        source = "parsed"
-
-                # Get raw text if needed
-                if text is None and search_mode in ("auto", "raw"):
-                    # Read raw content
-                    content = self.read(file_path)
-
-                    # Type narrowing: when return_metadata=False (default), result is bytes
-                    assert isinstance(content, bytes), "Expected bytes from read()"
-
-                    # Try to decode as text
-                    try:
-                        text = content.decode("utf-8")
-                        source = "raw"
-                    except UnicodeDecodeError:
-                        # Skip binary files
-                        continue
-
-                # Skip if no text available
-                if text is None:
-                    continue
-
-                # Search line by line
-                for line_num, line in enumerate(text.splitlines(), start=1):
-                    if len(results) >= max_results:
-                        break
-
-                    match = regex.search(line)
-                    if match:
-                        results.append(
-                            {
-                                "file": file_path,
-                                "line": line_num,
-                                "content": line,
-                                "match": match.group(0),
-                                "source": source,
-                            }
-                        )
-
-            except Exception:
-                # Skip files that can't be read
-                continue
-
-        return results
 
     # === Directory Operations ===
 
@@ -1535,7 +729,6 @@ class NexusFS(NexusFilesystem):
             )
             count = fs.export_metadata("backup.jsonl", filter=filter)
         """
-        import json
 
         output_file = Path(output_path)
         output_file.parent.mkdir(parents=True, exist_ok=True)
@@ -1682,7 +875,6 @@ class NexusFS(NexusFilesystem):
             result = fs.import_metadata("backup.jsonl", overwrite=True)
             # Returns ImportResult, but behaves like old (imported, skipped) tuple
         """
-        import json
 
         input_file = Path(input_path)
         if not input_file.exists():
@@ -1995,185 +1187,6 @@ class NexusFS(NexusFilesystem):
         """
         return self.metadata.batch_get_content_ids(paths)
 
-    # === Version Tracking (v0.3.5) ===
-
-    def get_version(self, path: str, version: int) -> bytes:
-        """Get a specific version of a file.
-
-        Retrieves the content for a specific version from CAS using the
-        version's content hash.
-
-        Args:
-            path: Virtual file path
-            version: Version number to retrieve
-
-        Returns:
-            File content as bytes for the specified version
-
-        Raises:
-            NexusFileNotFoundError: If file or version doesn't exist
-            InvalidPathError: If path is invalid
-
-        Example:
-            >>> # Get version 2 of a file
-            >>> content_v2 = nx.get_version("/workspace/data.txt", version=2)
-        """
-        path = self._validate_path(path)
-
-        # Get version metadata
-        version_meta = self.metadata.get_version(path, version)
-        if version_meta is None:
-            raise NexusFileNotFoundError(f"{path} (version {version})")
-
-        # Ensure version has content hash
-        if version_meta.etag is None:
-            raise NexusFileNotFoundError(f"{path} (version {version}) has no content")
-
-        # Read content from CAS using the version's content hash
-        route = self.router.route(
-            path,
-            tenant_id=self.tenant_id,
-            agent_id=self.agent_id,
-            is_admin=self.is_admin,
-            check_write=False,
-        )
-
-        content = route.backend.read_content(version_meta.etag)
-        return content
-
-    def list_versions(self, path: str) -> builtins.list[dict[str, Any]]:
-        """List all versions of a file.
-
-        Returns version history with metadata for each version.
-
-        Args:
-            path: Virtual file path
-
-        Returns:
-            List of version info dicts ordered by version number (newest first)
-
-        Raises:
-            InvalidPathError: If path is invalid
-
-        Example:
-            >>> versions = nx.list_versions("/workspace/SKILL.md")
-            >>> for v in versions:
-            ...     print(f"v{v['version']}: {v['size']} bytes, {v['created_at']}")
-        """
-        path = self._validate_path(path)
-        return self.metadata.list_versions(path)
-
-    def rollback(self, path: str, version: int, context: OperationContext | None = None) -> None:
-        """Rollback file to a previous version.
-
-        Updates the file to point to an older version's content from CAS.
-        Creates a new version entry marking this as a rollback.
-
-        Args:
-            path: Virtual file path
-            version: Version number to rollback to
-            context: Optional operation context for permission checks
-
-        Raises:
-            NexusFileNotFoundError: If file or version doesn't exist
-            InvalidPathError: If path is invalid
-            PermissionError: If user doesn't have write permission
-
-        Example:
-            >>> # Rollback to version 2
-            >>> nx.rollback("/workspace/data.txt", version=2)
-        """
-        path = self._validate_path(path)
-
-        # Check write permission
-        self._check_permission(path, Permission.WRITE, context)
-
-        # Route to backend
-        route = self.router.route(
-            path,
-            tenant_id=self.tenant_id,
-            agent_id=self.agent_id,
-            is_admin=self.is_admin,
-            check_write=True,
-        )
-
-        # Check readonly
-        if route.readonly:
-            raise PermissionError(f"Cannot rollback read-only path: {path}")
-
-        # Perform rollback in metadata store
-        self.metadata.rollback(path, version)
-
-        # Invalidate cache
-        if self.metadata._cache_enabled and self.metadata._cache:
-            self.metadata._cache.invalidate_path(path)
-
-    def diff_versions(
-        self, path: str, v1: int, v2: int, mode: str = "metadata"
-    ) -> dict[str, Any] | str:
-        """Compare two versions of a file.
-
-        Args:
-            path: Virtual file path
-            v1: First version number
-            v2: Second version number
-            mode: Diff mode - "metadata" (default) or "content"
-
-        Returns:
-            For "metadata" mode: Dict with metadata differences
-            For "content" mode: Unified diff string
-
-        Raises:
-            NexusFileNotFoundError: If file or version doesn't exist
-            InvalidPathError: If path is invalid
-            ValueError: If mode is invalid
-
-        Examples:
-            >>> # Get metadata diff
-            >>> diff = nx.diff_versions("/workspace/file.txt", v1=1, v2=3)
-            >>> print(f"Size changed: {diff['size_v1']} -> {diff['size_v2']}")
-
-            >>> # Get content diff
-            >>> diff_text = nx.diff_versions("/workspace/file.txt", v1=1, v2=3, mode="content")
-            >>> print(diff_text)
-        """
-        import difflib
-
-        path = self._validate_path(path)
-
-        if mode not in ("metadata", "content"):
-            raise ValueError(f"Invalid mode: {mode}. Must be 'metadata' or 'content'")
-
-        # Get metadata diff
-        meta_diff = self.metadata.get_version_diff(path, v1, v2)
-
-        if mode == "metadata":
-            return meta_diff
-
-        # Content diff mode
-        if not meta_diff["content_changed"]:
-            return "(no content changes)"
-
-        # Retrieve both versions' content
-        content1 = self.get_version(path, v1).decode("utf-8", errors="replace")
-        content2 = self.get_version(path, v2).decode("utf-8", errors="replace")
-
-        # Generate unified diff
-        lines1 = content1.splitlines(keepends=True)
-        lines2 = content2.splitlines(keepends=True)
-
-        diff_lines = list(
-            difflib.unified_diff(
-                lines1,
-                lines2,
-                fromfile=f"{path} (v{v1})",
-                tofile=f"{path} (v{v2})",
-                lineterm="",
-            )
-        )
-
-        return "\n".join(diff_lines)
-
     async def parse(
         self,
         path: str,
@@ -2241,183 +1254,6 @@ class NexusFS(NexusFilesystem):
             self.metadata.set_file_metadata(path, "parser_name", parser.name)
 
         return result
-
-    # === Permission Management (v0.3.0) ===
-
-    def chmod(
-        self,
-        path: str,
-        mode: int | str,
-        context: OperationContext | None = None,
-    ) -> None:
-        """Change file mode (permissions).
-
-        Requires the user to be the owner of the file or an admin.
-
-        Args:
-            path: Virtual file path
-            mode: Permission mode (int like 0o644 or string like '755')
-            context: Optional operation context (defaults to self._default_context)
-
-        Raises:
-            NexusFileNotFoundError: If file doesn't exist
-            InvalidPathError: If path is invalid
-            PermissionError: If user is not owner and not admin
-            ValueError: If mode is invalid
-
-        Examples:
-            >>> nx.chmod("/workspace/file.txt", 0o644)
-            >>> nx.chmod("/workspace/file.txt", "755")
-            >>> nx.chmod("/workspace/file.txt", "rwxr-xr-x")
-        """
-        from nexus.core.permissions import parse_mode
-
-        path = self._validate_path(path)
-
-        # Get file metadata
-        file_meta = self.metadata.get(path)
-        if not file_meta:
-            raise NexusFileNotFoundError(path)
-
-        # Get context (use default if not provided)
-        ctx = context or self._default_context
-
-        # Check if user is owner or admin
-        # Must be owner to chmod (unless admin or system)
-        if (
-            not ctx.is_admin
-            and not ctx.is_system
-            and file_meta.owner
-            and file_meta.owner != ctx.user
-        ):
-            raise PermissionError(
-                f"Access denied: Only the owner ('{file_meta.owner}') or admin "
-                f"can change permissions for '{path}'"
-            )
-
-        # Parse mode (handles int, octal string, or symbolic string)
-        if isinstance(mode, str):
-            mode_int = parse_mode(mode)
-        elif isinstance(mode, int):
-            mode_int = mode
-        else:
-            raise ValueError(f"mode must be int or str, got {type(mode)}")
-
-        # Update mode
-        file_meta.mode = mode_int
-        self.metadata.put(file_meta)
-
-        # Invalidate cache
-        if self.metadata._cache_enabled and self.metadata._cache:
-            self.metadata._cache.invalidate_path(path)
-
-    def chown(
-        self,
-        path: str,
-        owner: str,
-        context: OperationContext | None = None,
-    ) -> None:
-        """Change file owner.
-
-        Requires the user to be the current owner of the file or an admin.
-
-        Args:
-            path: Virtual file path
-            owner: New owner username
-            context: Optional operation context (defaults to self._default_context)
-
-        Raises:
-            NexusFileNotFoundError: If file doesn't exist
-            InvalidPathError: If path is invalid
-            PermissionError: If user is not owner and not admin
-
-        Examples:
-            >>> nx.chown("/workspace/file.txt", "alice")
-        """
-        path = self._validate_path(path)
-
-        # Get file metadata
-        file_meta = self.metadata.get(path)
-        if not file_meta:
-            raise NexusFileNotFoundError(path)
-
-        # Get context (use default if not provided)
-        ctx = context or self._default_context
-
-        # Check if user is owner or admin
-        # Must be owner to chown (unless admin or system)
-        if (
-            not ctx.is_admin
-            and not ctx.is_system
-            and file_meta.owner
-            and file_meta.owner != ctx.user
-        ):
-            raise PermissionError(
-                f"Access denied: Only the owner ('{file_meta.owner}') or admin "
-                f"can change ownership for '{path}'"
-            )
-
-        # Update owner
-        file_meta.owner = owner
-        self.metadata.put(file_meta)
-
-        # Invalidate cache
-        if self.metadata._cache_enabled and self.metadata._cache:
-            self.metadata._cache.invalidate_path(path)
-
-    def chgrp(
-        self,
-        path: str,
-        group: str,
-        context: OperationContext | None = None,
-    ) -> None:
-        """Change file group.
-
-        Requires the user to be the owner of the file or an admin.
-
-        Args:
-            path: Virtual file path
-            group: New group name
-            context: Optional operation context (defaults to self._default_context)
-
-        Raises:
-            NexusFileNotFoundError: If file doesn't exist
-            InvalidPathError: If path is invalid
-            PermissionError: If user is not owner and not admin
-
-        Examples:
-            >>> nx.chgrp("/workspace/file.txt", "developers")
-        """
-        path = self._validate_path(path)
-
-        # Get file metadata
-        file_meta = self.metadata.get(path)
-        if not file_meta:
-            raise NexusFileNotFoundError(path)
-
-        # Get context (use default if not provided)
-        ctx = context or self._default_context
-
-        # Check if user is owner or admin
-        # Must be owner to chgrp (unless admin or system)
-        if (
-            not ctx.is_admin
-            and not ctx.is_system
-            and file_meta.owner
-            and file_meta.owner != ctx.user
-        ):
-            raise PermissionError(
-                f"Access denied: Only the owner ('{file_meta.owner}') or admin "
-                f"can change group for '{path}'"
-            )
-
-        # Update group
-        file_meta.group = group
-        self.metadata.put(file_meta)
-
-        # Invalidate cache
-        if self.metadata._cache_enabled and self.metadata._cache:
-            self.metadata._cache.invalidate_path(path)
 
     def close(self) -> None:
         """Close the filesystem and release resources."""
